feat: refactor history section
