+++ src/components/sidebar-bar/history/history.tsx
@@ -1,0 +1,70 @@
+import React, { useMemo } from "react";
+import { subDays } from "date-fns";
+import { useChatHistoryStore } from "../../../store/chat-history-store";
+import { HistoryGroup } from "./history-group";
+
+const today = new Date();
+const sevenDaysAgo = subDays(today, 7);
+
+export const History: React.FC = () => {
+	const { chatHistory } = useChatHistoryStore();
+
+	const chatsToday = useMemo(
+		() =>
+			chatHistory.filter(
+				(chat) =>
+					new Date(chat.timestamp).toDateString() === today.toDateString(),
+			),
+		[chatHistory, today],
+	);
+
+	const chatsLastSevenDays = useMemo(
+		() =>
+			chatHistory.filter(
+				(chat) =>
+					new Date(chat.timestamp) >= sevenDaysAgo &&
+					!chatsToday.includes(chat),
+			),
+		[chatHistory, sevenDaysAgo, chatsToday],
+	);
+
+	const chatsOlderThanSevenDays = useMemo(
+		() =>
+			chatHistory.filter(
+				(chat) =>
+					!chatsLastSevenDays.includes(chat) && !chatsToday.includes(chat),
+			),
+		[chatHistory, chatsLastSevenDays, chatsToday],
+	);
+
+	const chatGroups = useMemo(() => {
+		const groups = [];
+
+		if (chatsToday.length > 0) {
+			groups.push({ label: "Heute", chats: chatsToday });
+		}
+
+		if (chatsLastSevenDays.length > 0) {
+			groups.push({ label: "Letzte 7 Tage", chats: chatsLastSevenDays });
+		}
+
+		if (chatsOlderThanSevenDays.length > 0) {
+			groups.push({
+				label: "Älter als 7 Tage",
+				chats: chatsOlderThanSevenDays,
+			});
+		}
+
+		return groups;
+	}, [chatsToday, chatsLastSevenDays, chatsOlderThanSevenDays]);
+
+	return (
+		<div>
+			<div className={`flex flex-col gap-4`}>
+				{chatGroups.map(({ label, chats }) => (
+					<HistoryGroup key={label} label={label} chats={chats} />
+				))}
+			</div>
+		</div>
+	);
+};
